move EncryptedKeyPairs.ts to the same place as other key pair type definitions
make signature a common attribute in all EncryptedKeyPairs

#tuta3620
