+++ src/platform-kit/crypto/index.ts
@@ -160,3 +160,4 @@ export * from "./instance-pipeline-crypto/SymmetricCipherFacade.js"
 export * from "./instance-pipeline-crypto/decryption/InstanceDecryptor.js"
 export * from "./instance-pipeline-crypto/encryption/SubKeyProvider.js"
 export { RsaImplementation } from "./encryption/RsaImplementation.js"
+export { EncryptedKeyPairs, EncryptedRsaX25519KeyPairs, EncryptedRsaKeyPairs, EncryptedPqKeyPairs } from "./encryption/EncryptedKeyPairs"
