+++ src/platform-kit/crypto/instance-pipeline-crypto/CryptoWrapper.ts
@@ -2,17 +2,7 @@ import { arrayEquals, stringToUtf8Uint8Array, Versioned } from "@tutao/utils"
 import { CryptoError } from "@tutao/crypto/error"
 import { Aes256Key, aes256RandomKey, AesKey, keyToUint8Array, uint8ArrayToKey } from "../encryption/symmetric/SymmetricCipherUtils.js"
 import { aesDecrypt, aesEncrypt } from "./Aes.js"
-import {
-	decryptKey,
-	decryptKeyPair,
-	EncryptedKeyPairs,
-	EncryptedPqKeyPairs,
-	EncryptedRsaKeyPairs,
-	EncryptedRsaX25519KeyPairs,
-	encryptKey,
-	encryptKyberKey,
-	encryptX25519Key,
-} from "./KeyEncryption.js"
+import { decryptKey, decryptKeyPair, encryptKey, encryptKyberKey, encryptX25519Key } from "./KeyEncryption.js"
 import { deriveX25519PublicKey, generateX25519KeyPair, X25519KeyPair, X25519PrivateKey, X25519PublicKey } from "../encryption/X25519.js"
 import { bytesToEd25519PrivateKey, Ed25519PrivateKey, ed25519PrivateKeyToBytes, Ed25519PublicKey, ed25519PublicKeyToBytes } from "../encryption/Ed25519.js"
 import {
@@ -32,6 +22,7 @@ import * as cryptoUtils from "../CryptoUtils.js"
 import { PQKeyPairs } from "../encryption/PQKeyPairs.js"
 import { hkdf } from "../hashes/HKDF.js"
 import { HkdfKeyDerivationDomains, VersionedEncryptedKey, VersionedKey } from "../CryptoTypes"
+import { EncryptedKeyPairs, EncryptedPqKeyPairs, EncryptedRsaKeyPairs, EncryptedRsaX25519KeyPairs } from "../encryption/EncryptedKeyPairs"
 
 type IdentityKeyPair = { privateEd25519Key: Uint8Array; identityKeyVersion: NumberString }
