+++ src/platform-kit/crypto/instance-pipeline-crypto/KeyEncryption.ts
@@ -10,44 +10,7 @@ import { Aes128Key, Aes256Key, AesKey } from "../encryption/symmetric/SymmetricC
 import { AesKeyLength, assert256BitKey, getKeyLengthInBytes } from "../encryption/symmetric/AesKeyLength.js"
 import { SYMMETRIC_CIPHER_FACADE } from "./SymmetricCipherFacade.js"
 import { ProgrammingError } from "@tutao/app-env"
-
-export abstract class EncryptedKeyPairs {
-	constructor() {} // public readonly pubKyberKey: null | Uint8Array, // public readonly pubEccKey: null | Uint8Array,
-}
-
-export class EncryptedPqKeyPairs extends EncryptedKeyPairs {
-	constructor(
-		public readonly pubEccKey: Uint8Array,
-		public readonly pubKyberKey: Uint8Array,
-		public readonly symEncPrivEccKey: Uint8Array,
-		public readonly symEncPrivKyberKey: Uint8Array,
-		public signature: null | object, //type PublicKeySignature not available in crypto package
-	) {
-		super()
-	}
-}
-
-export class EncryptedRsaKeyPairs extends EncryptedKeyPairs {
-	constructor(
-		public readonly pubRsaKey: Uint8Array,
-		public readonly symEncPrivRsaKey: Uint8Array,
-		public readonly signature: null | object, //type PublicKeySignature not available in crypto package
-	) {
-		super()
-	}
-}
-
-export class EncryptedRsaX25519KeyPairs extends EncryptedRsaKeyPairs {
-	constructor(
-		public readonly pubEccKey: Uint8Array,
-		pubRsaKey: Uint8Array,
-		public readonly symEncPrivEccKey: Uint8Array,
-		symEncPrivRsaKey: Uint8Array,
-		signature: null | object, //type PublicKeySignature not available in crypto package
-	) {
-		super(pubRsaKey, symEncPrivRsaKey, signature)
-	}
-}
+import { EncryptedKeyPairs, EncryptedPqKeyPairs, EncryptedRsaKeyPairs, EncryptedRsaX25519KeyPairs } from "../encryption/EncryptedKeyPairs"
 
 export function encryptKey(encryptionKey: AesKey, keyToBeEncrypted: AesKey): Uint8Array {
 	return SYMMETRIC_CIPHER_FACADE.encryptKey(encryptionKey, keyToBeEncrypted)
